Feat: added API for add to baske and get cart calls

+++ server/SPG/src/main/java/it/polito/SE2/P12/SPG/repository/ProductRepo.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ProductRepo
         extends JpaRepository<Product, Integer> {
-    Product findProductByProductId(Integer productId);
+    Product findProductByProductId(Long productId);
 }
